requirements.txt
absl-py==1.0.0
alabaster==0.7.12
alembic==1.7.5
arch==5.1.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
atari-py==0.2.6
attrs==21.4.0
autopage==0.4.0
Babel==2.9.1
backcall==0.2.0
black==21.12b0
bleach==4.1.0
box2d-py==2.3.8
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.9
click==7.1.2
cliff==3.10.0
cloudpickle==1.6.0
cmaes==0.8.2
cmd2==2.3.3
colorama==0.4.4
colorlog==6.6.0
colour==0.1.5
commonmark==0.9.1
coverage==6.2
cycler==0.11.0
DataProperty==0.54.2
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
docutils==0.17.1
dtreeviz==1.3.2
entrypoints==0.4
execnet==1.9.0
executing==0.8.3
feather-format==0.4.1
flake8==4.0.1
flake8-bugbear==21.11.29
fonttools==4.28.5
google-auth==2.3.3
google-auth-oauthlib==0.4.6
graphviz==0.19.1
greenlet==1.1.2
grpcio==1.43.0
gym==0.19.0
gym-minigrid==1.0.2
idna==3.3
imagesize==1.3.0
importlab==0.6.1
importlib-metadata==4.10.0
iniconfig==1.1.1
install==1.3.5
ipykernel==6.9.1
ipython==8.1.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
isort==5.10.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.0
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
jyp==0.1.3
kiwisolver==1.3.2
libcst==0.3.23
livereload==2.6.3
llvmlite==0.37.0
Mako==1.1.6
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mbstrdecoder==1.1.0
mccabe==0.6.1
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.11
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.6.3
ninja==1.10.2.3
notebook==6.4.8
numba==0.54.1
numpy==1.20.3
oauthlib==3.1.1
opencv-python==4.5.5.62
opencv-python-headless==4.5.5.62
optuna==2.10.0
packaging==21.3
panda-gym==1.1.1
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pathvalidate==2.5.0
patsy==0.5.2
pbr==5.8.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.0
#pkg_resources==0.0.0
platformdirs==2.4.1
plotly==5.5.0
pluggy==1.0.0
prettytable==2.5.0
prometheus-client==0.13.1
prompt-toolkit==3.0.28
property-cached==1.6.4
protobuf==3.19.1
psutil==5.9.0
psweep==0.5.1
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyaml==21.10.1
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybullet==3.2.1
pycodestyle==2.8.0
pycparser==2.21
pyenchant==3.2.2
pyflakes==2.4.0
pyglet==1.5.21
Pygments==2.11.1
pyparsing==3.0.6
pyperclip==1.8.2
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyrsistent==0.18.1
pytablewriter==0.64.1
pytest==6.2.5
pytest-cov==3.0.0
pytest-env==0.6.2
pytest-forked==1.4.0
pytest-xdist==2.5.0
python-dateutil==2.8.2
pytype==2021.12.15
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
requests==2.26.0
requests-oauthlib==1.3.0
rich==9.13.0
rliable==1.0.5
rsa==4.8
sb3-contrib==1.3.1a7
scikit-learn==1.0.2
scikit-obliquetree==0.1.4
scikit-optimize==0.9.0
scipy==1.7.3
seaborn==0.11.2
Send2Trash==1.8.0
shellingham==1.4.0
six==1.16.0
sklearn-sfa==0.1.4
snowballstemmer==2.2.0
Sphinx==4.3.2
sphinx-autobuild==2021.3.14
sphinx-autodoc-typehints==1.13.0
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sphinxcontrib-spelling==7.3.2
spyct-tstepi @ git+https://gitlab.com/TStepi/spyct.git@e4d83668dfef033cfd08a9efa6ec2e2a14547336
SQLAlchemy==1.4.29
stable-baselines3==1.4.0
stack-data==0.2.0
statsmodels==0.13.1
stevedore==3.5.0
tabledata==1.3.0
tabulate==0.8.9
tcolorpy==0.1.1
tenacity==8.0.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
terminado==0.13.1
testpath==0.6.0
threadpoolctl==3.0.0
toml==0.10.2
tomli==1.2.3
torch==1.10.1
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typepy==1.3.0
typer==0.3.2
typing-extensions==3.10.0.2
typing-inspect==0.7.1
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==3.5.2
zipp==3.7.0
